Adds bi-streaming demo
